file errors happening

+++ DecimalCalc.java
@@ -23,7 +23,7 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double subtract(Double n1, Double n2) 
     {
-        return Math.round(n1 - n2);
+        return (double) Math.round(n1 - n2);
     }
 
     /**
@@ -34,7 +34,7 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double multiply(Double n1, Double n2) 
     {
-        return Math.round(n1 * n2);
+        return (double) Math.round(n1 * n2);
     }
 
     /**
@@ -45,7 +45,7 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double divide(Double n1, Double n2) 
     {
-        return Math.round(n1 / n2);
+        return (double) Math.round(n1 / n2);
     }
 
     /**
@@ -56,7 +56,7 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double modulus(Double n1, Double n2) 
     {
-        return Math.round(n1 % n2);
+        return (double) Math.round(n1 % n2);
     }
 
     /**
@@ -67,7 +67,7 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double exponent(Double n1, Double n2) 
     {
-        return Math.round(Math.pow(n1, n2));
+        return (double) Math.round(Math.pow(n1, n2));
     }
 
     /**
@@ -77,7 +77,7 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double squareRoot(Double n1) 
     {
-        return Math.round(Math.sqrt(n1));
+        return (double) Math.round(Math.sqrt(n1));
     }
 
     /**
@@ -87,6 +87,6 @@ public class DecimalCalc implements CalculatorOperations<Double>
     @Override
     public Double cubeRoot(Double n1) 
     {
-        return Math.round(Math.cbrt(n1));
+        return (double) Math.round(Math.cbrt(n1));
     }
 }
